requirements.txt
urllib3<2
requests
pandas
python-dotenv